cleanup and fix fmax runner
workaround for toml bug? where inline tables are returned as
unserializable TomlDecoder... instances
