+++ xeda/plugins/lwc/flows/__init__.py
@@ -1,2 +1,3 @@
+from .power import VivadoPower as VivadoPowerLwc # alias for old name
 from .power import *
 from .timing_and_verification import *
